# src/main/java/org/example/chatapp/dto/AppInfo.java
package org.example.chatapp.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AppInfo {
    private String appName;
    private String developers;
    private String description;
}