# src/main/java/org/example/chatapp/dto/GetChats.java
package org.example.chatapp.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class GetChats {
    private List<GetChat> chats;
}