# src/main/java/org/example/chatapp/dto/NewMessage.java
package org.example.chatapp.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class NewMessage {
    @JsonProperty("chat_id")
    private String chatId;
    private String text;
}