# src/main/java/org/example/chatapp/dto/ReplyMessage.java
package org.example.chatapp.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ReplyMessage {
    private String text;
    private String createdAt;
    private String author;
}